pyproject.toml
[project]
name = "heliostat"
version = "0.1.0"
description = "Build custom rocks for sunbeam from ppas"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "msgspec>=0.20.0",
    "rich>=14.3.1",
    "ruamel-yaml>=0.19.1",
    "typer>=0.21.1",
    "xdg-base-dirs>=6.0.2",
]

[project.scripts]
heliostat = "heliostat.cli:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "ruff>=0.14.14",
]
